Connecting movie data to movie grid component
Connected the placeholder movie grid cards to endpoints to load real movie data, including the title, vote_average and release_date.

Created color-coordinated rating system, where color of the rating container changes based on the rating.

Set the featured slider up with a placeholder image to prepare for the incoming backdrop data that we are adding into our database into a new field, which is using the TMDB API to fetch the images.