Hero Video Implementation
Added a 'Netflix-style' video trailer in the background in place of the image in the Featured Slider!

Currently only a hardcoded link, will need to make use of the YouTube Player API to have more control over the start and end of the video that will be looped, instead of the entire trailer.

NOT muted by default, mute browser tab if using, I wanted to make sure the sound was also working :)