Changing Hero Video using Youtube Player API
The YouTube Player API is now implemented to display the video. This takes in the movie id from the video_url link, where the video key is extracted using regular expression.

The implementation involves handling instances of the YouTube player and making sure the cleanup is robust each time it unmounts when the featured slider changes movies.

A custom snippet loop also decides how long the video is and where it starts.