Movie details + Lightweight Endpoint
Made a more lightweight endpoint for the astro static rendering of paths. This endpoint retrieves only the ids instead of all the data, made specifically for pre-rendering static paths.

Movie details page is updated, styled and formatted with more data and information. (Popularity Bar utils didn't work properly for unknown reasons so the function is written again in the [id].astro page which works)

Lowered the default volume for featured slider hero video, could potentially add a volume slider.

Adjusted the movie card styling to make sure most of the titles in the posters fit within the card.