Created cinemind_toolkit.ipynb
A more streamlined and formatted python notebook which will be used in future to interact with the database instead of the api_and_db_testing.

The code to update the database will be added once all the additional data fields are finalised, as we're still developing and figuring out what additional data we need.