Reviews Section Completed
Added the reviews into the database model and endpoint and parsed the reviews data in the MovieReviews component.

This string is processed into an array of objects, with each review object having an author and content property.

These objects are used to display the reviews in a front end visualisation with an optional focus mode, letting the user navigate through the reviews with forward and back arrow buttons.

The focus mode brings up a modal container that shows the review in a larger view which can be exited out of via a cross button.