Fixed Movie, Backdrop and Video Implementation
After adding multiple links instead of just one link to posters, backdrops and trailers, the implementation of how the urls were read had to be changed.

I parsed these strings by splitting them on the first comma, and trimmed the rest of the string. This would cut down the string to just the first link.