Search Bar Functionality Complete
Fixed the search bar functionality to update the movie grid.

Before, the grid would dynamically update everytime the search term changes which is very expensive in terms of API calls. This is now changed to only fetching the results after hitting the submit button, which saves a lot of API calls.