Updated Toolkit + Prep for new db population
Updated the toolkit with new functions to match the new database schema as well as connecting to the new migration-based database. New functions are also added for clearing the entire database data as well as deleting a specific movie and deleting orphaned records.

This update preps the migration database for data population from the API